requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
black==25.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.0
cron-descriptor==1.4.5
cryptography==44.0.3
Django==5.2.1
django-axes==8.0.0
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-extensions==4.1
django-guardian==3.0.0
django-prometheus==2.3.1
django-simple-history==3.8.0
django-storages==1.14.6
django-timezone-field==7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.1.0
flake8==7.2.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
kombu==5.5.3
mccabe==0.7.0
msgpack==1.1.0
mypy==1.15.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
py==1.11.0
pycodestyle==2.13.0
pycparser==2.22
pydot==4.0.0
pyflakes==3.3.2
PyJWT==2.9.0
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.3
sentry-sdk==2.28.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
